Struct buf_list::BufList

pub struct BufList { /* private fields */ }

Data composed of a list of Bytes chunks.

For more, see the crate documentation.

Implementations

impl BufList

pub fn new() -> Self

Creates a new, empty, BufList.

pub fn with_capacity(capacity: usize) -> Self

Creates a new, empty, BufList with the given capacity.

pub fn num_chunks(&self) -> usize

Returns the total number of chunks in this BufList.

Examples

use buf_list::BufList;

let buf_list = vec![&b"hello"[..], &b"world"[..]].into_iter().collect::<BufList>();
assert_eq!(buf_list.num_chunks(), 2);

pub fn num_bytes(&self) -> usize

Returns the total number of bytes across all chunks.

Examples

use buf_list::BufList;

let buf_list = vec![&b"hello"[..], &b"world"[..]].into_iter().collect::<BufList>();
assert_eq!(buf_list.num_bytes(), 10);

pub fn get_chunk(&self, index: usize) -> Option<&Bytes>

Provides a reference to the chunk at the given index.

Examples

use buf_list::BufList;
use bytes::Bytes;

let buf_list = vec![&b"hello"[..], &b"world"[..]].into_iter().collect::<BufList>();
assert_eq!(buf_list.get_chunk(1), Some(&Bytes::from(&b"world"[..])));

pub fn iter(&self) -> Iter<'_>

Iterates over the chunks in this list.

pub fn push_chunk<B: Buf>(&mut self, data: B) -> Bytes

Adds a new chunk to this list.

If the provided Buf is zero-length, it will not be added to the list.

Examples

use buf_list::BufList;
use bytes::{Buf, Bytes};

let mut buf_list = BufList::new();

// &'static [u8] implements Buf.
buf_list.push_chunk(&b"hello"[..]);
assert_eq!(buf_list.chunk(), &b"hello"[..]);

// Bytes also implements Buf.
buf_list.push_chunk(Bytes::from_static(&b"world"[..]));
assert_eq!(buf_list.num_chunks(), 2);

// A zero-length `Buf` will not be added to the list.
buf_list.push_chunk(Bytes::new());
assert_eq!(buf_list.num_chunks(), 2);