buf_list

Struct Cursor

pub struct Cursor<T> { /* private fields */ }

A Cursor wraps an in-memory BufList and provides it with a Seek implementation.

Cursors allow BufLists to implement Read and BufRead, allowing a BufList to be used anywhere you might use a reader or writer that does actual I/O.

The cursor may either own or borrow a BufList: both Cursor<BufList> and Cursor<&BufList> are supported.

Optional features

tokio1: With this feature enabled, Cursor implements the tokio crate’s AsyncSeek, AsyncRead and AsyncBufRead.

futures03: With this feature enabled, Cursor implements the futures crate’s AsyncSeek, AsyncRead and AsyncBufRead.

Implementations

impl<T: AsRef<BufList>> Cursor<T>

pub fn new(inner: T) -> Cursor<T>

Creates a new cursor wrapping the provided BufList.

Examples

use buf_list::{BufList, Cursor};

let cursor = Cursor::new(BufList::new());

pub fn into_inner(self) -> T

Consumes this cursor, returning the underlying value.

Examples

use buf_list::{BufList, Cursor};

let cursor = Cursor::new(BufList::new());

let vec = cursor.into_inner();

pub const fn get_ref(&self) -> &T

Gets a reference to the underlying value in this cursor.

Examples

use buf_list::{BufList, Cursor};

let cursor = Cursor::new(BufList::new());

let reference = cursor.get_ref();

pub const fn position(&self) -> u64

Returns the current position of this cursor.

Examples

use buf_list::{BufList, Cursor};
use std::io::prelude::*;
use std::io::SeekFrom;

let mut cursor = Cursor::new(BufList::from(&[1, 2, 3, 4, 5][..]));

assert_eq!(cursor.position(), 0);

cursor.seek(SeekFrom::Current(2)).unwrap();
assert_eq!(cursor.position(), 2);

cursor.seek(SeekFrom::Current(-1)).unwrap();
assert_eq!(cursor.position(), 1);

pub fn set_position(&mut self, pos: u64)

Sets the position of this cursor.

Examples

use buf_list::{BufList, Cursor};

let mut cursor = Cursor::new(BufList::from(&[1, 2, 3, 4, 5][..]));

assert_eq!(cursor.position(), 0);

cursor.set_position(2);
assert_eq!(cursor.position(), 2);

cursor.set_position(4);
assert_eq!(cursor.position(), 4);